+++ lib/lightning_web/live/credential_live/credential_form_component.ex
@@ -127,7 +127,7 @@ defmodule LightningWeb.CredentialLive.CredentialFormComponent do
        changeset: changeset,
        update_body: update_body,
        projects: projects,
-       selected_oauth_client: assigns.credential.oauth_client,
+       selected_oauth_client: assigns.oauth_client,
        schema: schema,
        selected_project: nil,
        selected_projects: selected_projects,
@@ -501,7 +501,6 @@ defmodule LightningWeb.CredentialLive.CredentialFormComponent do
                   phx-target={@myself}
                   id={"salesforce_sandbox_instance_checkbox_#{@credential.id || "new"}"}
                 />
-
                 <.input
                   :if={@schema in ["salesforce_oauth"]}
                   type="text"