+++ test/lightning_web/live/credential_live_test.exs
@@ -868,12 +868,20 @@ defmodule LightningWeb.CredentialLiveTest do
            user: user,
            token: token
          } do
+      oauth_client = insert(:oauth_client, user: user, userinfo_endpoint: nil)
+
       credential =
         insert(:credential,
           user: user,
           schema: "oauth",
-          body: token,
-          oauth_client: build(:oauth_client, user: user, userinfo_endpoint: nil)
+          body: %{"apiVersion" => "v1"},
+          oauth_token:
+            build(:oauth_token,
+              body: token,
+              user: user,
+              oauth_client: oauth_client
+            ),
+          oauth_client: oauth_client
         )
 
       Mox.stub(Lightning.AuthProviders.OauthHTTPClient.Mock, :call, fn env,
@@ -1190,7 +1198,7 @@ defmodule LightningWeb.CredentialLiveTest do
              |> Enum.find(fn pc -> pc.project_id == project_2.id end)
 
       token =
-        Lightning.AuthProviders.Common.TokenBody.new(credential.body)
+        Lightning.AuthProviders.Common.TokenBody.new(credential.oauth_token.body)
 
       assert %{
                access_token: "ya29.a0AVvZ",
@@ -1209,12 +1217,25 @@ defmodule LightningWeb.CredentialLiveTest do
 
       oauth_client = insert(:oauth_client, user: user)
 
+      oauth_token = %{
+        "access_token" => "test_access_token",
+        "refresh_token" => "test_refresh_token",
+        "token_type" => "bearer",
+        "expires_in" => 3600
+      }
+
       credential =
         insert(:credential,
           name: "OAuth credential",
           oauth_client: oauth_client,
           user: user,
           schema: "oauth",
+          oauth_token:
+            build(:oauth_token,
+              body: oauth_token,
+              user: user,
+              oauth_client: oauth_client
+            ),
           project_credentials: [
             %{project: project_1},
             %{project: project_2},
@@ -1311,25 +1332,26 @@ defmodule LightningWeb.CredentialLiveTest do
         insert(:credential,
           name: "my-credential",
           schema: "oauth",
-          body: %{
-            "access_token" => "access_token",
-            "refresh_token" => "refresh_token",
-            "expires_at" =>
-              Timex.now() |> Timex.shift(days: 4) |> DateTime.to_unix(),
-            "scope" =>
-              String.split(oauth_client.mandatory_scopes, ",") |> Enum.join(" ")
-          },
+          body: %{"apiVersion" => "v1"},
+          oauth_token:
+            build(:oauth_token,
+              body: %{
+                "access_token" => "access_token",
+                "refresh_token" => "refresh_token",
+                "expires_at" =>
+                  Timex.now() |> Timex.shift(days: 4) |> DateTime.to_unix(),
+                "scope" =>
+                  String.split(oauth_client.mandatory_scopes, ",")
+                  |> Enum.join(" ")
+              },
+              user: user,
+              oauth_client: oauth_client
+            ),
           user: user,
           oauth_client: oauth_client
         )
 
       {:ok, index_live, _html} = live(conn, ~p"/credentials")
-
-      index_live |> select_credential_type(oauth_client.id)
-      index_live |> click_continue()
-
-      assert index_live
-             |> has_element?("#scope_selection_new")
 
       refute index_live |> has_element?("#re-authorize-banner")
 
@@ -1354,17 +1376,24 @@ defmodule LightningWeb.CredentialLiveTest do
       expires_at = DateTime.to_unix(DateTime.utc_now()) + 3600
 
       credential =
-        credential_fixture(
-          user_id: user.id,
+        insert(:credential,
+          user: user,
           schema: "oauth",
-          body: %{
-            access_token: "ya29.a0AVvZ...",
-            refresh_token: "1//03vpp6Li...",
-            expires_at: expires_at,
-            scope:
-              String.split(oauth_client.mandatory_scopes, ",") |> Enum.join(" ")
-          },
-          oauth_client_id: oauth_client.id
+          body: %{"apiVersion" => "v1"},
+          oauth_token:
+            build(:oauth_token,
+              body: %{
+                "access_token" => "ya29.a0AVvZ...",
+                "refresh_token" => "1//03vpp6Li...",
+                "expires_at" => expires_at,
+                "scope" =>
+                  String.split(oauth_client.mandatory_scopes, ",")
+                  |> Enum.join(" ")
+              },
+              user: user,
+              oauth_client: oauth_client
+            ),
+          oauth_client: oauth_client
         )
 
       {:ok, edit_live, _html} = live(conn, ~p"/credentials")
@@ -1390,16 +1419,23 @@ defmodule LightningWeb.CredentialLiveTest do
       expires_at = DateTime.to_unix(DateTime.utc_now()) - 50
 
       credential =
-        credential_fixture(
-          user_id: user.id,
+        insert(:credential,
+          user: user,
           schema: "oauth",
-          body: %{
-            access_token: "ya29.a0AVvZ...",
-            refresh_token: "1//03vpp6Li...",
-            expires_at: expires_at,
-            scope:
-              String.split(oauth_client.mandatory_scopes, ",") |> Enum.join(" ")
-          },
+          body: %{"apiVersion" => "v1"},
+          oauth_token:
+            build(:oauth_token,
+              body: %{
+                access_token: "ya29.a0AVvZ...",
+                refresh_token: "1//03vpp6Li...",
+                expires_at: expires_at,
+                scope:
+                  String.split(oauth_client.mandatory_scopes, ",")
+                  |> Enum.join(" ")
+              },
+              user: user,
+              oauth_client: oauth_client
+            ),
           oauth_client_id: oauth_client.id
         )
 
@@ -1428,7 +1464,18 @@ defmodule LightningWeb.CredentialLiveTest do
           name: "My OAuth Credential",
           schema: "oauth",
           oauth_client: oauth_client,
-          user: user
+          user: user,
+          oauth_token:
+            build(:oauth_token,
+              body: %{
+                "access_token" => "test_access_token",
+                "refresh_token" => "test_refresh_token",
+                "token_type" => "bearer",
+                "expires_in" => 3600
+              },
+              user: user,
+              oauth_client: oauth_client
+            )
         )
 
       {:ok, view, html} = live(conn, ~p"/credentials")
@@ -1441,6 +1488,7 @@ defmodule LightningWeb.CredentialLiveTest do
                "span[phx-hook='Tooltip', aria-label='OAuth client not found']"
              )
 
+      # Now it's safe to delete the oauth client
       Repo.delete!(oauth_client)
 
       {:ok, view, html} = live(conn, ~p"/credentials")
@@ -1603,14 +1651,15 @@ defmodule LightningWeb.CredentialLiveTest do
       credential =
         Lightning.Credentials.list_credentials(user) |> List.first()
 
+      assert credential.body == %{"apiVersion" => "34"}
+
       assert %{
                access_token: "ya29.a0AVvZ",
                refresh_token: "1//03vpp6Li",
                expires_at: 3600,
-               scope: "scope1 scope2",
-               apiVersion: "34"
+               scope: "scope1 scope2"
              } =
-               credential.body
+               credential.oauth_token.body
                |> Map.new(fn {k, v} -> {String.to_atom(k), v} end)
     end
   end